refactor mdl
