+++ src/main/java/org/dbos/apiary/benchmarks/RetroBenchmark.java
@@ -5,8 +5,8 @@ import org.dbos.apiary.client.ApiaryWorkerClient;
 import org.dbos.apiary.function.FunctionOutput;
 import org.dbos.apiary.function.ProvenanceBuffer;
 import org.dbos.apiary.postgres.PostgresConnection;
-import org.dbos.apiary.procedures.postgres.replay.*;
-import org.dbos.apiary.procedures.postgres.retro.MDLSubscribeTxn;
+import org.dbos.apiary.procedures.postgres.moodle.*;
+import org.dbos.apiary.procedures.postgres.moodle.MDLSubscribeTxn;
 import org.dbos.apiary.utilities.ApiaryConfig;
 import org.dbos.apiary.worker.ApiaryNaiveScheduler;
 import org.dbos.apiary.worker.ApiaryWorker;
