+++ src/test/java/org/dbos/apiary/ProvenanceTests.java
@@ -4,10 +4,10 @@ import com.google.protobuf.InvalidProtocolBufferException;
 import org.dbos.apiary.client.ApiaryWorkerClient;
 import org.dbos.apiary.function.ProvenanceBuffer;
 import org.dbos.apiary.postgres.PostgresConnection;
-import org.dbos.apiary.procedures.postgres.replay.MDLFetchSubscribers;
-import org.dbos.apiary.procedures.postgres.replay.MDLForumInsert;
-import org.dbos.apiary.procedures.postgres.replay.MDLIsSubscribed;
-import org.dbos.apiary.procedures.postgres.retro.MDLSubscribeTxn;
+import org.dbos.apiary.procedures.postgres.moodle.MDLFetchSubscribers;
+import org.dbos.apiary.procedures.postgres.moodle.MDLForumInsert;
+import org.dbos.apiary.procedures.postgres.moodle.MDLIsSubscribed;
+import org.dbos.apiary.procedures.postgres.moodle.MDLSubscribeTxn;
 import org.dbos.apiary.procedures.postgres.tests.PostgresProvenanceBasic;
 import org.dbos.apiary.procedures.postgres.tests.PostgresProvenanceJoins;
 import org.dbos.apiary.procedures.postgres.tests.PostgresProvenanceMultiRows;
